finished refactoring api routes
