+++ server/src/routes/auth/index.ts
@@ -1,0 +1,10 @@
+import { Router } from "express";
+import authUserEmailRoute from "./email";
+
+const authRouter = Router();
+
+
+authRouter.use(authUserEmailRoute);
+
+
+export default authRouter;
